Flesh out failing test for AuctionMessageTranslator.process_message/1
The book includes an unused `chat` argument, but they specifically say
“as it happens, we don’t need one for the current functionality”, so I’m
calling Yagni.

Rather than mocking an interface (behaviour), this version is expecting
a message to be sent to a process. In real life this will be the main
application server, but for the test it will be to the test process
itself, allowing us to simply use `assert_receive`.